requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.0.3
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pluggy==1.5.0
pypdf==5.1.0
pytesseract==0.3.13
pytest==8.3.3
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
tzdata==2024.2
Werkzeug==3.1.3
